All doubles changed to float for better optimization

+++ src/main/java/net/smokeybbq/bittermelon/medical/simulation/IVAdministration.java
@@ -7,7 +7,7 @@ import net.smokeybbq.bittermelon.medical.simulation.compartments.SimpleCompartme
 
 public class IVAdministration extends PBPKModel {
 
-    public IVAdministration(double dosage, Character character, Substance drug) {
+    public IVAdministration(float dosage, Character character, Substance drug) {
         super(dosage, character, drug);
         simpleCompartments = new SimpleCompartment[]{GI, liver, kidney, lung, heart, brain, adiposeTissue, bone, muscle, lymphatic, endocrine, other};
     }
@@ -20,9 +20,9 @@ public class IVAdministration extends PBPKModel {
 
     @Override
     public void simulation() {
-        double circulatoryConcentration = circulatory.getConcentration(drug);
+        float circulatoryConcentration = circulatory.getConcentration(drug);
 
-        double circulatoryDerivative = circulatory.getDerivative(circulatoryConcentration, simpleCompartments, drug);
+        float circulatoryDerivative = circulatory.getDerivative(circulatoryConcentration, simpleCompartments, drug);
 
         handleSimpleCompartments();
         circulatory.updateConcentration(drug, circulatoryDerivative, timeStep);
